Move server index to src/index
