--- src/server/index.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const server = express();
-const routes = require('../routes');
-
-server.use(express.json());
-server.use('/api', routes);
-
-module.exports = server;
