+++ tests/routes.test.js
@@ -1,10 +1,10 @@
-const server = require('../src/server');
+const server = require('../src');
 const request = require('supertest');
 
 describe('Test public routes', () =>
 {
-    it("should respond with a 200 response and a 'Hello World!' body in /api route", async () =>
+    it("should respond with a 200 response and a 'CRUD Microservice Server' body in / route", async () =>
     {
-        return request(server).get('/api').expect(200, 'Hello World!');
+        return request(server).get('/').expect(200, 'CRUD Microservice Server');
     });
 });
